Move crash tests to ui tests
