--- tests/crashes/133966.rs
@@ -1,3 +1,0 @@
-//@ known-bug: #133966
-pub struct Data([[&'static str]; 5_i32]);
-const _: &'static Data = unsafe { &*(&[] as *const Data) };
