--- tests/crashes/136416.rs
@@ -1,6 +1,0 @@
-//@ known-bug: #136416
-#![feature(generic_const_exprs)]
-struct State<const S : usize = {}> where[(); S] :;
-
-struct Foo;
-struct State2<const S: usize = Foo> where [(); S]:;
