--- tests/crashes/94846.rs
@@ -1,6 +1,0 @@
-//@ known-bug: #94846
-#![feature(generic_const_exprs)]
-
-struct S<const C:() = {}>() where S<{}>:;
-
-pub fn main() {}
