+++ tests/ui/const-generics/generic_const_exprs/enum-const-generic-discriminant.stderr
@@ -1,0 +1,10 @@
+error: generic parameters may not be used in enum discriminant values
+  --> $DIR/enum-const-generic-discriminant.rs:14:15
+   |
+LL |     A = t::<{ N }>() as u8,
+   |               ^ cannot perform const operation using `N`
+   |
+   = note: const parameters may not be used in enum discriminant values
+
+error: aborting due to 1 previous error
+
