+++ tests/ui/consts/const-eval/133966.stderr
@@ -1,0 +1,15 @@
+error[E0308]: mismatched types
+  --> $DIR/133966.rs:2:34
+   |
+LL | pub struct Data([[&'static str]; 5_i32]);
+   |                                  ^^^^^ expected `usize`, found `i32`
+   |
+help: change the type of the numeric literal from `i32` to `usize`
+   |
+LL - pub struct Data([[&'static str]; 5_i32]);
+LL + pub struct Data([[&'static str]; 5_usize]);
+   |
+
+error: aborting due to 1 previous error
+
+For more information about this error, try `rustc --explain E0308`.
